minor tweak IUrl
