--- app/definitions/IUrl.ts
@@ -1,6 +1,0 @@
-export interface IUrl {
-	title: string;
-	description: string;
-	image: string;
-	url: string;
-}
